configure start script

+++ start.sh
@@ -16,10 +16,22 @@ nodeSetup(){
 nginxSetup(){
     echo '############################## Setting up pm2, nginx and getting the repo ########################################'
     sudo npm install pm2 -g -y
-    git clone https://github.com/babadee001/DevOps-AWS-Intro
+    if [ -d HBClient ]; then
+        echo '####################### Removing existing folder ######################'
+        sudo rm -rf HBClient
+    fi
+    git clone https://github.com/babadee001/HBClient
     sudo apt-get install nginx
-    cd /etc/nginx/sites-enabled/default
-    sudo rm /etc/nginx/sites-enabled/default
+    sudo /etc/init.d/nginx start
+    if [ -d /etc/nginx/sites-enabled/default ]; then
+        sudo rm /etc/nginx/sites-enabled/default
+    fi
+    if [ -d /etc/nginx/sites-enabled/HelloBooks ]; then
+        echo ' ######################## Removing existing config for app ##############################'
+        sudo rm -rf /etc/nginx/sites-available/HelloBooks
+        sudo rm -rf /etc/nginx/sites-enabled/HelloBooks
+    fi
+
     sudo bash -c 'cat > /etc/nginx/sites-available/HelloBooks <<EOF
     server {
             listen 80;
@@ -27,16 +39,14 @@ nginxSetup(){
             location / {
                     proxy_pass 'http://$ip:8000';
             }
-    }
-    EOF'
+    }'
     sudo ln -s /etc/nginx/sites-available/HelloBooks /etc/nginx/sites-enabled/HelloBooks
-    sudo /etc/init.d/nginx start
-    sudo /etc/init.d/nginx restart
+    sudo service nginx restart
 }
 
 installDep(){
     echo '######################### Installing App Dependencies ##################################'
-    cd DevOps-AWS-Intro
+    cd HBClient
     sudo npm install --unsafe-perm
 }
 
@@ -53,8 +63,11 @@ startPm2() {
      pm2 startOrRestart ecosystem.config.js
 }
 
-nodeSetup
-nginxSetup
-installDep
-sslSetup
-startPm2
+main(){
+    nodeSetup
+    nginxSetup
+    installDep
+    sslSetup
+    startPm2
+}
+main
